*add 3d lut color grading
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1163 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/client/ref.h
@@ -313,7 +313,7 @@ typedef struct image_s {
 	qboolean	hasHandle;
 	
 	//lut description
-	uint		lutSize;
+	float		lutSize;
 	char		*lutName;
 } image_t;
